+++ json_serializable/lib/src/type_helpers/config_types.dart
@@ -45,6 +45,7 @@ class ClassConfig {
   final String constructor;
   final bool createFactory;
   final bool createToJson;
+  final bool createFieldMap;
   final bool disallowUnrecognizedKeys;
   final bool explicitToJson;
   final FieldRename fieldRename;
@@ -60,6 +61,7 @@ class ClassConfig {
     required this.constructor,
     required this.createFactory,
     required this.createToJson,
+    required this.createFieldMap,
     required this.disallowUnrecognizedKeys,
     required this.explicitToJson,
     required this.fieldRename,
@@ -76,6 +78,8 @@ class ClassConfig {
         checked: config.checked ?? ClassConfig.defaults.checked,
         anyMap: config.anyMap ?? ClassConfig.defaults.anyMap,
         constructor: config.constructor ?? ClassConfig.defaults.constructor,
+        createFieldMap:
+            config.createFieldMap ?? ClassConfig.defaults.createFieldMap,
         createFactory:
             config.createFactory ?? ClassConfig.defaults.createFactory,
         createToJson: config.createToJson ?? ClassConfig.defaults.createToJson,
@@ -101,6 +105,7 @@ class ClassConfig {
     constructor: '',
     createFactory: true,
     createToJson: true,
+    createFieldMap: false,
     disallowUnrecognizedKeys: false,
     explicitToJson: false,
     fieldRename: FieldRename.none,
@@ -115,6 +120,7 @@ class ClassConfig {
         constructor: constructor,
         createFactory: createFactory,
         createToJson: createToJson,
+        createFieldMap: createFieldMap,
         ignoreUnannotated: ignoreUnannotated,
         explicitToJson: explicitToJson,
         includeIfNull: includeIfNull,