+++ json_serializable/test/config_test.dart
@@ -126,6 +126,10 @@ void main() {
             lastLine = "type 'int' is not a subtype of type 'String?' in type "
                 'cast';
             break;
+          case 'create_to_json':
+            lastLine = "type 'int' is not a subtype of type 'bool?' in type "
+                'cast';
+            break;
           default:
             lastLine =
                 "type 'int' is not a subtype of type 'bool?' in type cast";
@@ -152,6 +156,7 @@ const _invalidConfig = {
   'checked': 42,
   'constructor': 42,
   'create_factory': 42,
+  'create_field_map': 42,
   'create_to_json': 42,
   'disallow_unrecognized_keys': 42,
   'explicit_to_json': 42,