+++ json_serializable/test/test_sources/test_sources.dart
@@ -15,6 +15,7 @@ class ConfigurationImplicitDefaults {
   constructor: '',
   createFactory: true,
   createToJson: true,
+  createFieldMap: false,
   disallowUnrecognizedKeys: false,
   explicitToJson: false,
   fieldRename: FieldRename.none,